Update RN from 0.59 to 0.61.5 for IOS
Sets the xcode project with the right configuration, eliminating
the library manual linking inside the project and delegate it to
the auto linking from RN0.61.5 version.

Some android files are uploaded as well to help in the Android
configuration.

This change is necessary because when submitting to the Appstore
an email is received saying that Apple do not accept apps with
UIWebView. It happens that this is mainly generated by the WebView
component from React in RN0.59. In >= RN 0.60 there is no such
component, therefore the warnings disappear.

>= RN0.60 uses the Webview component from the community,
which translates into different internal IOS Webview component
that Apple is not against to.