Allow navigation to Stores when update required
Implement UpdateApp component UI with ability to navigate to stores,
and edit App.js structure.

Reason for these changes is that the Android Native version did force
users to update to newer versions, therefore there was no or small
segmentation of app versions across the user base. In React Native
this update enforcement was not implemented and new segmentation
started to show up for Android users.