Prevent app to crash when description/s and title not set
Events objects have description/s and title props for internationalization
purposes, some old events do not have this new props and what happens
is that the app crashes because it can not find such props.

To prevent that a fallback logic is implemented to detect if props exist or not
and replace the value expected from then by legacy props from events.