Add Remote configuration for Discord links
Add Firebase Remote configuration code to dynamically handle remote
objects and variables into the application.

Some times Discord links in the app expired and to fix this a new build
was needed to be published which means hours of having a broken
link inside the app. Now a 12h gap is used for fixing bad links from
discord. 12h is the default time Firebase Remote Config stablishes
to prevent DB throtelling (check documentation).

Remote config can be used in the future for an extended variety
of use cases which brings flexibility to the app and shortens
iterations. Be aware that Firebase remote config does not work
well with styles.css files.